Move lib/builtin.js to lib/builtin/ and split it
Split the internal helpers used by the compiled code (lib/builtin/ops.js),
the class definitions of TT values (lib/builtin/values.js) and the
type definitions of TT operators (lib/builtin/defs.js).
