+++ lib/ast.js
@@ -22,7 +22,7 @@ module.exports.FunctionDef = FunctionDef;
 module.exports.ExpressionSignature = ExpressionSignature;
 module.exports.ImportStmt = ImportStmt;
 
-const builtin = require('./builtin_values');
+const builtin = require('./builtin/values');
 
 adt.nativeClone = function nativeClone(x) {
     if (x === null || x === undefined)
