+++ lib/ast_api.js
@@ -13,7 +13,7 @@ const assert = require('assert');
 
 const Ast = require('./ast');
 const Type = require('./type');
-const Builtin = require('./builtin');
+const Builtin = require('./builtin/defs');
 const { prettyprint, prettyprintExample } = require('./prettyprint');
 
 const { isUnaryStreamToStreamOp,
