+++ lib/builtin/ops.js
@@ -2,7 +2,7 @@
 //
 // This file is part of ThingTalk
 //
-// Copyright 2015-2016 The Board of Trustees of the Leland Stanford Junior University
+// Copyright 2015-2018 The Board of Trustees of the Leland Stanford Junior University
 //
 // Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
 //
@@ -11,17 +11,17 @@
 
 const assert = require('assert');
 
-const Type = require('./type');
-const Ast = require('./ast');
+const Ast = require('../ast');
 const AsyncQueue = require('consumer-queue');
 
-// we split the module so that AST can also load it, because
-// we need Ast to define function def at the end of the file
-module.exports = require('./builtin_values');
-const Location = module.exports.Location;
-const Entity = module.exports.Entity;
-const Time = module.exports.Time;
-const Currency = module.exports.Currency;
+const BuiltinValues = require('./values');
+const Location = BuiltinValues.Location;
+const Entity = BuiltinValues.Entity;
+const Time = BuiltinValues.Time;
+const Currency = BuiltinValues.Currency;
+
+// Implementations of the ThingTalk operators, and library helpers used by the
+// compiled TT code
 
 function arrayEquals(a, b) {
     if (a.length !== b.length)
@@ -306,198 +306,7 @@ function combineOutputTypes(t1, t2) {
 }
 module.exports.combineOutputTypes = combineOutputTypes;
 
-module.exports.BinaryOps = {
-    '>': {
-        types: [[Type.String, Type.String, Type.Boolean],
-                [Type.Measure(''), Type.Measure(''), Type.Boolean],
-                [Type.Number, Type.Number, Type.Boolean],
-                [Type.Date, Type.Date, Type.Boolean],
-                [Type.Time, Type.Time, Type.Boolean],
-                [Type.Currency, Type.Currency, Type.Boolean]],
-        op: '>'
-    },
-    '<': {
-        types: [[Type.String, Type.String, Type.Boolean],
-                [Type.Measure(''), Type.Measure(''), Type.Boolean],
-                [Type.Number, Type.Number, Type.Boolean],
-                [Type.Date, Type.Date, Type.Boolean],
-                [Type.Time, Type.Time, Type.Boolean],
-                [Type.Currency, Type.Currency, Type.Boolean]],
-        op: '<'
-    },
-    '>=': {
-        types: [[Type.String, Type.String, Type.Boolean],
-                [Type.Measure(''), Type.Measure(''), Type.Boolean],
-                [Type.Number, Type.Number, Type.Boolean],
-                [Type.Date, Type.Date, Type.Boolean],
-                [Type.Time, Type.Time, Type.Boolean],
-                [Type.Currency, Type.Currency, Type.Boolean]],
-        op: '>='
-    },
-    '<=': {
-        types: [[Type.String, Type.String, Type.Boolean],
-                [Type.Measure(''), Type.Measure(''), Type.Boolean],
-                [Type.Number, Type.Number, Type.Boolean],
-                [Type.Date, Type.Date, Type.Boolean],
-                [Type.Time, Type.Time, Type.Boolean],
-                [Type.Currency, Type.Currency, Type.Boolean]],
-        op: '<='
-    },
-    '==': {
-        types: [['a', 'a', Type.Boolean]],
-        fn: 'equality',
-    },
-    '=~': {
-        types: [[Type.String, Type.String, Type.Boolean]],
-        fn: 'like'
-    },
-    '~=': {
-        types: [[Type.String, Type.String, Type.Boolean]],
-        fn: 'like',
-        flip: true
-    },
-    starts_with: {
-        types: [[Type.String, Type.String, Type.Boolean]],
-        fn: 'startsWith',
-    },
-    ends_with: {
-        types: [[Type.String, Type.String, Type.Boolean]],
-        fn: 'endsWith',
-    },
-    prefix_of: {
-        types: [[Type.String, Type.String, Type.Boolean]],
-        fn: 'startsWith',
-        flip: true
-    },
-    suffix_of: {
-        types: [[Type.String, Type.String, Type.Boolean]],
-        fn: 'endsWith',
-        flip: true
-    },
-    'contains': {
-        types: [[Type.Array('a'), 'a', Type.Boolean]],
-        fn: 'contains',
-    },
-    'in_array': {
-        types: [['a', Type.Array('a'), Type.Boolean]],
-        fn: 'contains',
-        flip: true
-    },
-    'has_member': {
-        types: [[Type.Entity('tt:contact_group'), Type.Entity('tt:contact'), Type.Boolean]],
-    },
-    'group_member': {
-        types: [[Type.Entity('tt:contact'), Type.Entity('tt:contact_group'), Type.Boolean]],
-    }
-};
-
 function getTime(d) {
     return new Time(d.getHours(), d.getMinutes(), d.getSeconds());
 }
 module.exports.getTime = getTime;
-
-module.exports.UnaryOps = {
-    '!': {
-        types: [[Type.Boolean, Type.Boolean]],
-        op: '!'
-    },
-    'get_time': {
-        types: [[Type.Date, Type.Time]],
-        fn: 'getTime'
-    }
-};
-
-module.exports.ScalarExpressionOps = {
-    '+': {
-        types: [[Type.String, Type.String, Type.String],
-                [Type.Number, Type.Number, Type.Number],
-                [Type.Currency, Type.Currency, Type.Currency],
-                [Type.Measure(''), Type.Measure(''), Type.Measure('')],
-                [Type.Date, Type.Measure('ms'), Type.Date],
-                [Type.Time, Type.Measure('ms'), Type.Time],
-                [Type.Measure('ms'), Type.Date, Type.Date],
-                [Type.Measure('ms'), Type.Time, Type.Time]],
-        op: '+'
-    },
-    '-': {
-        types: [[Type.Number, Type.Number, Type.Number],
-                [Type.Currency, Type.Currency, Type.Currency],
-                [Type.Measure(''), Type.Measure(''), Type.Measure('')],
-                [Type.Date, Type.Measure('ms'), Type.Date],
-                [Type.Time, Type.Measure('ms'), Type.Time]],
-        op: '-'
-    },
-    '*': {
-        types: [[Type.Number, Type.Number, Type.Number],
-                [Type.Currency, Type.Number, Type.Currency],
-                [Type.Measure(''), Type.Number, Type.Measure('')]],
-        op: '*'
-    },
-    '/': {
-        types: [[Type.Number, Type.Number, Type.Number],
-                [Type.Currency, Type.Number, Type.Currency],
-                [Type.Measure(''), Type.Number, Type.Measure('')]],
-        op: '/'
-    },
-    '%': {
-        types: [[Type.Number, Type.Number, Type.Number]],
-        op: '%'
-    },
-    '**': {
-        types: [[Type.Number, Type.Number, Type.Number]],
-        op: '**'
-    },
-    'distance': {
-        types: [[Type.Location, Type.Location, Type.Measure('m')]],
-        op: '~'
-    }
-};
-
-module.exports.Aggregations = {
-    'max': {
-        types: [[Type.Number, Type.Number],
-                [Type.Currency, Type.Currency],
-                [Type.Measure(''), Type.Measure('')]]
-    },
-    'min': {
-        types: [[Type.Number, Type.Number],
-                [Type.Currency, Type.Currency],
-                [Type.Measure(''), Type.Measure('')]]
-    },
-    'sum': {
-        types: [[Type.Number, Type.Number],
-                [Type.Currency, Type.Currency],
-                [Type.Measure(''), Type.Measure('')]]
-    },
-    'avg': {
-        types: [[Type.Number, Type.Number],
-                [Type.Currency, Type.Currency],
-                [Type.Measure(''), Type.Measure('')]]
-    },
-    'count': {
-        types: [[Type.Any, Type.Number]]
-    }
-};
-
-module.exports.ArgMinMax = {
-    'argmax': {
-        types: [Type.Number, Type.Measure(''), Type.Currency]
-    },
-    'argmin': {
-        types: [Type.Number, Type.Measure(''), Type.Currency]
-    }
-};
-
-function builtinFunction(name) {
-    return new Ast.FunctionDef('action', name, [], false, false, {}, {});
-}
-module.exports.emptyFunction = builtinFunction;
-module.exports.Triggers = {
-};
-module.exports.Actions = {
-    'notify': builtinFunction('notify'),
-    'return': builtinFunction('return'),
-    'save': builtinFunction('save'),
-};
-module.exports.Queries = {
-};
