+++ lib/builtin/values.js
@@ -9,7 +9,7 @@
 // See COPYING for details
 "use strict";
 
-const { stringEscape } = require('./escaping');
+const { stringEscape } = require('../escaping');
 
 class Entity {
     constructor(id, display) {
@@ -85,11 +85,11 @@ class Currency {
         this.value = value;
         this.code = code;
     }
-    
+
     valueOf() {
         return this.value;
     }
-    
+
     toString() {
         return `${this.value} ${this.code.toUpperCase()}`;
     }
@@ -99,7 +99,7 @@ class Currency {
         options.currency = this.code.toUpperCase();
         return this.value.toLocaleString(locale, options);
     }
-    
+
     toJSSource() {
         return `new __builtin.Currency(${this.value}, "${this.code}")`;
     }
