+++ lib/compiler/ops-to-jsir.js
@@ -13,7 +13,7 @@ const assert = require('assert');
 
 const Ast = require('../ast');
 const Type = require('../type');
-const Builtin = require('../builtin');
+const Builtin = require('../builtin/defs');
 const { NotImplementedError } = require('../errors');
 
 const JSIr = require('./jsir');
