+++ lib/describe.js
@@ -15,7 +15,7 @@ const Ast = require('./ast');
 const Type = require('./type');
 const { clean } = require('./utils');
 const FormatUtils = require('./format_utils');
-const { Currency } = require('./builtin_values');
+const { Currency } = require('./builtin/values');
 
 class Describer {
     constructor(gettext, locale, timezone) {
