+++ lib/permission_checker.js
@@ -14,7 +14,8 @@ const smt = require('smtlib');
 
 const Ast = require('./ast');
 const Type = require('./type');
-const Builtin = require('./builtin');
+const BuiltinDefs = require('./builtin/defs');
+const BuiltinOps = require('./builtin/ops');
 
 function arrayEquals(a, b) {
     if (a === null && b === null)
@@ -784,9 +785,9 @@ function evaluateOp(builtinOp, arg1, arg2) {
     if (builtinOp.op)
         return OP_FUNCTIONS[builtinOp.op](arg1, arg2);
     else if (builtinOp.flip)
-        return Builtin[builtinOp.fn](arg2, arg1);
+        return BuiltinOps[builtinOp.fn](arg2, arg1);
     else
-        return Builtin[builtinOp.fn](arg1, arg2);
+        return BuiltinOps[builtinOp.fn](arg1, arg2);
 }
 
 function isRemoteSend(fn) {
@@ -997,7 +998,7 @@ class RuleTransformer {
                 if (filter.operator === 'group_member')
                     result = isGroupMember(jslhs, jsrhs, groupmap);
                 else
-                    result = evaluateOp(Builtin.BinaryOps[filter.operator], jslhs, jsrhs);
+                    result = evaluateOp(BuiltinDefs.BinaryOps[filter.operator], jslhs, jsrhs);
                 if (result === true)
                     return Ast.BooleanExpression.True;
                 else if (result === false)
