+++ lib/typecheck.js
@@ -14,7 +14,7 @@ const assert = require('assert');
 const Ast = require('./ast');
 const Type = require('./type');
 const Utils = require('./utils');
-const Builtin = require('./builtin');
+const Builtin = require('./builtin/defs');
 const Units = require('./units');
 
 const ALLOWED_PRINCIPAL_TYPES = new Set([
