+++ test/test_builtin_values.js
@@ -2,8 +2,8 @@
 
 const assert = require('assert');
 
-const __builtin = require('../lib/builtin_values');
-const { equality } = require('../lib/builtin');
+const __builtin = require('../lib/builtin/values');
+const { equality } = require('../lib/builtin/ops');
 
 function testValueOf(what, expected) {
     assert.strictEqual(+what, expected);
